+++ ui/src/components/timeline/useTimelineChartColors.ts
@@ -15,8 +15,6 @@ const DATAZOOM_HANDLE_OPACITY = 0.3;
 const DATAZOOM_FILLER_OPACITY = 0.2;
 const DATAZOOM_EMPHASIS_HANDLE_OPACITY = 0.5;
 const DATAZOOM_LABEL_BACKGROUND_OPACITY = 0.5;
-const OVERLAY_OPACITY = 0.9;
-const OVERLAY_OPACITY_DARK = 0.6;
 const OVERLAY_LIGHTEN = 0.6;
 const OVERLAY_LIGHTEN_DARK = 0.4;
 
@@ -62,7 +60,6 @@ export function useTimelineChartColors() {
         DATAZOOM_EMPHASIS_HANDLE_OPACITY
       ),
 
-      overlayOpacity: theme === THEME_DARK ? OVERLAY_OPACITY_DARK : OVERLAY_OPACITY,
       overlayLighten: theme === THEME_DARK ? OVERLAY_LIGHTEN_DARK : OVERLAY_LIGHTEN,
     };
   }, [theme]);
